handle issue with multiline registry string
There was an issue with handling multiline registry key strings that contained line feeds (\n). They would cause the ini parser to throw an error.  Need to enclose the value in double quotes before applying universal newlines in order to flag the parser that it needs to process it as a multiline value.